+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * [EACQAPW-4452] App Based Flow Payment support
 * [EACQAPW-4453] Update Third Party bundled xcframeworks
 * [EACQAPW-5460] ImageLoader Tests
+* [EACQAPW-5472] TDSWrapperBuilder Tests
 
 ### Fixed